// src/app/modules/marketing/model/timeline.model.ts
export interface ITimelineMarketingModel {
    Id: number
    Subject: string
    Description: string
    StartTime: string
    EndTime: string
    TopicId: number
    TopicName: string
    CustomerId: number
    Icon: string
    CategoryColor: string
    Latitude: string
    Longitude: string
    PIC: string
}